Final : Beta 1
Need a little more tuning to get a better output

+++ src/main/scala-2.10/doDecisionTrees.scala
@@ -1,4 +1,3 @@
-import java.util.Calendar
 import org.apache.spark.mllib.linalg.Vectors
 import org.apache.spark.mllib.regression.LabeledPoint
 import org.apache.spark.mllib.tree.DecisionTree
@@ -6,8 +5,7 @@ import org.apache.spark.mllib.tree.model.DecisionTreeModel
 import org.apache.spark.rdd.RDD
 import org.apache.spark.sql.DataFrame
 import org.joda.time
-import org.joda.time.{Interval, DateTime}
-import org.joda.time.base.AbstractInterval
+import org.joda.time.{ DateTime}
 
 /**
  * Created by sneha on 12/5/2015.
@@ -42,7 +40,7 @@ object doDecisionTrees {
           (model.predict(Vectors.dense(math.round((l(1).toString.toDouble) * 10),
             math.round(l(2).toString.toDouble * 10),
             l(3).toString.toDouble,
-            math.round(l(4).toString.toDouble),
+            math.round(l(4).toString.toInt.toDouble),
 
             math.round(l(5).toString.toDouble))).toInt), l(7).toString))
       else (l(0).toString, 0, (l(7).toString.toInt).toString))
